backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.7
bcrypt==4.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
spacy==3.7.2
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
pylint==2.17.5
gunicorn==21.2.0
SQLAlchemy==2.0.21
magic-pdf[full]
Pillow==10.0.1
PyPDF2==3.0.1
pdfplumber
pandas==2.1.1
openpyxl==3.1.2
flask-socketio==5.3.6
python-socketio==5.9.0
marshmallow==3.20.1
language-tool-python==2.7.1
google-generativeai==0.3.2
aiohttp==3.8.5
tenacity==8.2.3
